PATCH - sticky
Press MODKEY+s (default) to make a client ‘sticky’. A sticky client is visible on all tags. This is similar to setting the client’s tags to all 1’s, but with the ability to easily return it to its original tag by toggling it off by pressing MODKEY+s again.

https://dwm.suckless.org/patches/sticky/