PATCH - tab
Transforms the monocle layout into a ‘'tabbed’‘ layout if more than one window is present on the monocle view. Navigating from window to window is done by clicking on the window tabs or using the usual Mod1-j, Mod1-k keys. The tabs are arranged in a bar on top or at bottom of the screen, which can also be displayed in the other layouts than monocle. Three display modes can be selected at run time, auto display, permanent display and no display. In permanent mode the tab bar is always display independently of the layout, while in the auto mode it is displayed only with the monocle layout and in presence of several windows.

https://dwm.suckless.org/patches/tab/